requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.5
psycopg2-binary==2.9.9
python-dateutil==2.8.2
